+++ Changelog.md
@@ -2,6 +2,6 @@
 
 ## 2024-10-11
 
--   Change log to tracing crate
--   Add tracing subscriber for debug
--   Add dotenv load and .env.template for example
+- Change log to tracing crate
+- Add tracing subscriber for debug
+- Add dotenv load and .env.template for example